move Controls into layers.py
